The express matrix had a bug - a int() needed to be a ceil(). But seeing this made me realise that express could be a floating point rather than integer - then all the difficulties disappear. However, it does now fail some unit tests. At the same time, I added charge injection line mode, as a new express matrix.
